+++ docs/conf.py
@@ -50,7 +50,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'ModernGL'
-copyright = '2018, Szabolcs Dombi'
+copyright = '2020, Szabolcs Dombi'
 author = 'Szabolcs Dombi'
 
 # The version info for the project you're documenting, acts as replacement for
@@ -98,6 +98,8 @@ html_theme = 'sphinx_rtd_theme'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+html_logo = "_static/logo.png"
+html_favicon = "_static/favicon.ico"
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.