+++ moderngl/compute_shader.py
@@ -1,4 +1,4 @@
-from typing import Tuple, Union
+from typing import Generator, Tuple, Union
 
 from .program_members import (Attribute, Subroutine, Uniform, UniformBlock,
                               Varying)
@@ -8,8 +8,15 @@ __all__ = ['ComputeShader']
 
 class ComputeShader:
     '''
-        A Compute Shader is a Shader Stage that is used entirely for computing arbitrary information.
-        While it can do rendering, it is generally used for tasks not directly related to drawing.
+        A Compute Shader is a Shader Stage that is used entirely for computing
+        arbitrary information. While it can do rendering, it is generally used
+        for tasks not directly related to drawing.
+
+        - Compute shaders support uniforms are other member object just like a
+          :py:class:`moderngl.Program`.
+        - Storage buffers can be bound using :py:meth:`Buffer.bind_to_storage_buffer`.
+        - Uniform buffers can be bound using :py:meth:`Buffer.bind_to_uniform_block`.
+        - Images can be bound using :py:meth:`Texture.bind_to_image`.
     '''
 
     __slots__ = ['mglo', '_members', '_glo', 'ctx', 'extra']
@@ -26,21 +33,57 @@ class ComputeShader:
         return '<ComputeShader: %d>' % self.glo
 
     def __eq__(self, other):
+        """Compares to compute shaders ensuring the internal opengl name/id is the same"""
         return type(self) is type(other) and self.mglo is other.mglo
 
     def __getitem__(self, key) -> Union[Uniform, UniformBlock, Subroutine, Attribute, Varying]:
+        """Get a member such as uniforms, uniform blocks, subroutines,
+        attributes and varyings by name.
+
+        .. code-block:: python
+
+            # Get a uniform
+            uniform = program['color']
+
+            # Uniform values can be set on the returned object
+            # or the `__setitem__` shortcut can be used.
+            program['color'].value = 1.0, 1.0, 1.0, 1.0
+
+            # Still when writing byte data we need to use the `write()` method
+            program['color'].write(buffer)
+        """
         return self._members[key]
 
-    def __iter__(self):
+    def __setitem__(self, key, value):
+        """Set a value of uniform or uniform block
+
+        .. code-block:: python
+
+            # Set a vec4 uniform
+            uniform['color'] = 1.0, 1.0, 1.0, 1.0
+
+            # Optionally we can store references to a member and set the value directly
+            uniform = program['color']
+            uniform.value = 1.0, 0.0, 0.0, 0.0
+
+            uniform = program['cameraMatrix']
+            uniform.write(camera_matrix)
+        """
+        self._members[key].value = value
+
+    def __iter__(self) -> Generator[str, None, None]:
+        """Yields the internal members names as strings.
+        This includes all members such as uniforms, attributes etc.
+        """
         yield from self._members
 
-    @property
-    def source(self) -> str:
-        '''
-            str: The source code of the compute shader.
-        '''
+    # @property
+    # def source(self) -> str:
+    #     '''
+    #         str: The source code of the compute shader.
+    #     '''
 
-        return self.mglo.source
+    #     return self.mglo.source
 
     @property
     def glo(self) -> int: