+++ tests/test_context.py
@@ -1,0 +1,75 @@
+from unittest import TestCase
+import moderngl
+import numpy
+import platform
+
+
+class ContextTests(TestCase):
+
+    def test_create_destroy(self):
+        """Create and destroy a context"""
+        for _ in range(25):
+            ctx = moderngl.create_context(standalone=True)
+            ctx.release()
+
+    def test_context_switch(self):
+        """Ensure context switching is working"""
+        ctx1 = moderngl.create_context(standalone=True)
+        ctx2 = moderngl.create_context(standalone=True)
+
+        with ctx1 as ctx:
+            buffer1 = ctx.buffer(reserve=1024)
+        
+        with ctx2 as ctx:
+            buffer2 = ctx.buffer(reserve=1024)
+
+        self.assertEqual(buffer1.glo, buffer2.glo)
+        ctx1.release()
+        ctx2.release()
+
+    def test_exit(self):
+        """Ensure the previous context was activated on exit"""
+        ctx1 = moderngl.create_context(standalone=True)
+        ctx2 = moderngl.create_context(standalone=True)
+
+        with ctx1 as ctx:
+            ctx.buffer(reserve=1024)
+
+        # Will error out if no context is active "moderngl.error.Error: cannot create buffer"
+        ctx1.buffer(reserve=1024)
+
+        ctx1.release()
+        ctx2.release()
+
+    def test_share(self):
+        """Create resources with shared context"""
+        if platform.system().lower() in ["darwin", "linux"]:
+            self.skipTest('Context sharing not supported on darwin')
+
+        data1 = numpy.array([1, 2, 3, 4], dtype='u1')
+        data2 = numpy.array([4, 3, 2, 1], dtype='u1')
+
+        ctx1 = moderngl.create_context(standalone=True)
+        ctx2 = moderngl.create_context(standalone=True, share=True)
+
+        with ctx1 as ctx:
+            b1 = ctx.buffer(data=data1)
+
+        with ctx2 as ctx:
+            b2 = ctx.buffer(data=data2)
+
+        # Because the resources are shared the name should increment
+        self.assertEqual(b1.glo, 1)
+        self.assertEqual(b2.glo, 2)
+
+        # Ensure we can read the same buffer data in both contexts
+        with ctx1:
+            self.assertEqual(b1.read(), b'\x01\x02\x03\x04')
+            self.assertEqual(b2.read(), b'\x04\x03\x02\x01')
+
+        with ctx2:
+            self.assertEqual(b1.read(), b'\x01\x02\x03\x04')
+            self.assertEqual(b2.read(), b'\x04\x03\x02\x01')
+
+        ctx1.release()
+        ctx2.release()